refactor: one charged-total helper for ledger and InferenceCall rows
Live aggregate-budget charges and durable usage persistence now share
provider_usage::{charged_usage_total, persisted_usage_counts}, matching
Lean AggregateBudget.Usage.chargedTotal so ATIF sums cannot diverge
from the in-process ledger. Also finish a leftover validate_agent_request
rename from the 1087 stack rebase.
